Remove agency creation from AdminAgencyManagementService
