+++ Api/Services/Agents/IAgencyService.cs
@@ -8,6 +8,8 @@ namespace HappyTravel.Edo.Api.Services.Agents
 {
     public interface IAgencyService
     {
+        Task<Result<AgencyInfo>> Create(RegistrationAgencyInfo agencyInfo, int counterpartyId, int? parentAgencyId);
+        
         Task<Result<SlimAgencyInfo>> Get(AgentContext agent, string languageCode = LocalizationHelper.DefaultLanguageCode);
 
         Task<Result<SlimAgencyInfo>> Edit(AgentContext agent, EditAgencyRequest editAgencyRequest,
