+++ HappyTravel.Edo.UnitTests/Utility/AdministratorServicesMockCreationHelper.cs
@@ -3,9 +3,7 @@ using System.Threading.Tasks;
 using CSharpFunctionalExtensions;
 using HappyTravel.Edo.Api.AdministratorServices;
 using HappyTravel.Edo.Api.Infrastructure;
-using HappyTravel.Edo.Api.Infrastructure.Options;
 using HappyTravel.Edo.Api.NotificationCenter.Services;
-using HappyTravel.Edo.Api.Services.Locations;
 using HappyTravel.Edo.Api.Services.Management;
 using HappyTravel.Edo.Api.Services.Payments.Accounts;
 using HappyTravel.Edo.Common.Enums;
@@ -16,7 +14,6 @@ using HappyTravel.Edo.Data.Payments;
 using HappyTravel.Edo.UnitTests.Mocks;
 using HappyTravel.Money.Enums;
 using Microsoft.EntityFrameworkCore.Infrastructure;
-using Microsoft.Extensions.Options;
 using Moq;
 
 namespace HappyTravel.Edo.UnitTests.Utility
@@ -52,13 +49,8 @@ namespace HappyTravel.Edo.UnitTests.Utility
         }
 
 
-        public AdminAgencyManagementService GetAgencyManagementService(EdoContext context)
-        {
-            return new(context,
-                Mock.Of<IDateTimeProvider>(),
-                Mock.Of<IManagementAuditService>(),
-                Mock.Of<ILocalityInfoService>());
-        }
+        public AdminAgencyManagementService GetAgencyManagementService(EdoContext context) 
+            => new(context, Mock.Of<IDateTimeProvider>(), Mock.Of<IManagementAuditService>());
 
 
         public AgencyVerificationService GetAgencyVerificationService(EdoContext context)
